+ Changing all references from uint64_t to unsigned long long in c4.
The reason for this is due to a standards clash best seen in MPI_Traits.hh.
Although we'd like to insist on uint64_t (ANSI C 99), we are only guaranteed
unsigned long long in the MPI standard. So fundamentally we have to pass using
the trait from the following function in the MPI traits:

template<>
struct MPI_Traits<unsigned long long>
{
    static MPI_Datatype element_type() { return MPI_UNSIGNED_LONG_LONG; }
};

We can only rely on short <= int <= long <= long long -- hopefully
int64_t fits into one of those boxes in a parallel call.

+++ src/c4/Compare.cc
@@ -76,7 +76,8 @@ bool check_global_equiv(int local_value)
 
 //---------------------------------------------------------------------------//
 /*!
- * \brief Function to check the equivalence of an uint64_t across all processors.
+ * \brief Function to check the equivalence of an unsigned long long
+ *        across all processors.
  *
  * This function is (hopefully) a temporary parallel check function that more
  * properly belongs in C4.  It is used to check the equivalence of a given
@@ -86,8 +87,7 @@ bool check_global_equiv(int local_value)
  * \param local_value integer value to check against
  * \return true if equivalent across all processors; false if not
  */
-//bool check_global_equiv(unsigned long long local_value)
-bool check_global_equiv(uint64_t local_value)
+bool check_global_equiv(unsigned long long local_value)
 {
     
     const int node  = rtt_c4::node();
