+++ src/c4/test/tstCompare.cc
@@ -20,7 +20,6 @@
 #include <iostream>
 #include <vector>
 #include <cmath>
-#include <stdint.h>
 
 using namespace std;
 using namespace rtt_c4;
@@ -136,8 +135,8 @@ int main(int argc, char *argv[])
         test_equivalence(10, 11);           // int
         test_equivalence(10.0001, 11.0001); // double
         test_equivalence(10.0001, 10.0002); // double
-        test_equivalence(uint64_t(10000000000),
-                         uint64_t(200000000000)); // uint64_t
+        test_equivalence(static_cast<unsigned long long>(10000000000),
+                         static_cast<unsigned long long>(200000000000));
 
     }
     catch (exception &err)
